Add --cache option to use obs_scm service clones
The obs_scm service caches git clones and reuses them, greatly
reducing waiting times and bandwidth usage among other things.
Unfortunately it only exposes to the source services (or build
services) the checkout without the .git metadata. It also does
not support the most common Debian workflow, which does not store
the .dsc in the repository, given it's the output of a build.

If a cached repository is present, clone from it rather than from the
URL. Git is clever and uses hard links when possible so it's fast.

This is done by using obs_scm Python libraries to avoid reinventing
the wheel, although technically it's not a public library so the API
could change.